version 9
Se realizan correcciones, comentado con JavaDoc y se implementa el service y el controllador de Rol

+++ src/main/java/com/SAFE_Rescue/API_Administrador/repository/RolRepository.java
@@ -4,7 +4,11 @@ import com.SAFE_Rescue.API_Administrador.modelo.Rol;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+/**
+ * Repositorio para la gestión de roles
+ * Maneja operaciones CRUD desde la base de datos usando Jakarta
+ */
 @Repository
-public interface RolRepository extends JpaRepository<Rol, Integer> {
+public interface RolRepository extends JpaRepository<Rol, Long> {
 
 }
